data-generator/requirements.txt
kafka-python==2.0.2
uuid==1.30